+++ v2/Source/System/IO/eosStreams.cpp
@@ -14,8 +14,8 @@ export namespace eos {
     	public:
     		virtual ~Stream() = default;
 
-    		virtual ResultU32 write(const uint8_t *buffer, uint32_t length) = 0;
-    		virtual ResultU32 read(uint8_t *buffer, uint32_t bufferSize) = 0;
+    		virtual ResultU32 write(const uint8_t *buffer, size_t length) = 0;
+    		virtual ResultU32 read(uint8_t *buffer, size_t bufferSize) = 0;
     };
 
 }
